ml/CodeOptimizer/requirements.txt
anthropic==0.13.0
pylint==3.1.0
autopep8==2.1.0
python-dotenv==1.0.1
langchain-community==0.1.5
langchain-text-splitters==0.1.2